requirements.txt
# ==========================================
# AI Agents & LLM Core
# ==========================================
autogen-agentchat==0.7.5
autogen-core==0.7.5
openai==2.14.0
transformers==4.57.3
sentence-transformers==5.2.0
chromadb==1.3.5
modelscope==1.33.0
huggingface-hub==0.36.0
tiktoken==0.12.0
tokenizers==0.22.1
accelerate==1.12.0

# ==========================================
# Computer Vision & OCR
# ==========================================
ultralytics==8.3.241
paddleocr==3.3.2
paddlex==3.3.12
opencv-python==4.12.0.88
opencv-contrib-python==4.10.0.84
pillow==11.1.0
albumentations==2.0.8
scikit-image==0.25.2

# ==========================================
# Document & PDF Processing
# ==========================================
magic-pdf==1.3.12
mineru==2.6.8
PyMuPDF==1.24.14
pdfminer.six==20250506
pdf2image==1.17.0
pdftext==0.6.3
pypdf==6.5.0

# ==========================================
# Data Science & Machine Learning
# ==========================================
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
numpy==2.0.1
pandas==2.3.3
scipy==1.15.3
scikit-learn==1.7.2
matplotlib==3.10.8
seaborn==0.13.2
networkx==3.4.2
faiss-cpu==1.13.1

# ==========================================
# Web, API & Networking
# ==========================================
fastapi==0.127.0
uvicorn==0.40.0
gradio==5.49.1
requests==2.32.5
httpx==0.28.1
beautifulsoup4==4.14.3
duckduckgo_search==8.1.1
neo4j==6.0.3
boto3==1.42.15

# ==========================================
# Infrastructure & Utilities
# ==========================================
docker==7.1.0
kubernetes==34.1.0
pydantic==2.11.10
pydantic-settings==2.12.0
python-dotenv==1.2.1
loguru==0.7.3
tqdm==4.67.1
click==8.3.1
pyyaml==6.0.3

# ==========================================
# General Dependencies
# ==========================================
aiofiles==24.1.0
aistudio-sdk==0.3.8
albucore==0.0.24
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
asttokens==3.0.0
attrs==25.4.0
av==16.0.1
backoff==2.2.1
bce-python-sdk==0.9.56
bcrypt==5.0.0
botocore==1.42.16
brotli==1.2.0
brotlicffi==1.2.0.0
build==1.3.0
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chonky==0.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
ddgs==9.10.0
debugpy==1.8.16
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
doclayout_yolo==0.0.4
durationpy==0.10
easydict==1.13
exceptiongroup==1.3.0
executing==2.2.1
fake-useragent==2.2.0
fast-langdetect==0.2.5
fasttext-predict==0.9.2.4
ffmpy==1.0.0
filelock==3.20.0
FLAML==2.3.6
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
ftfy==6.3.1
future==1.0.0
gmpy2==2.2.2
google-auth==2.43.0
googleapis-common-protos==1.72.0
gradio_client==1.13.3
gradio_pdf==0.0.22
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx-retries==0.4.5
huggingface==0.0.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.31.0
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
json_repair==0.54.3
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lxml==6.0.2
magika==0.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mineru_vl_utils==0.1.19.1
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mmh3==5.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
parso==0.8.5
platformdirs==4.5.0
polars==1.36.1
polars-runtime-32==1.36.1
posthog==5.4.0
prettytable==3.17.0
primp==0.15.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyautogen==0.4
pybase64==1.4.3
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
pyzmq==27.1.0
qwen-vl-utils==0.0.14
rapid-table==1.0.5
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.7
requests-oauthlib==2.0.0
rich==14.2.0
robust-downloader==0.0.2
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
ruff==0.14.10
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
semantic-version==2.10.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.1
stack_data==0.6.3
starlette==0.50.0
stringzilla==4.5.1
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.3.0
tomlkit==0.13.3
tornado==6.5.4
traitlets==5.14.3
typer==0.21.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
ujson==5.11.0
ultralytics-thop==2.0.18
urllib3==2.6.1
watchfiles==1.1.1
wcwidth==0.2.13
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
zipp==3.23.0

# ==========================================
# WIN
# ==========================================
pywin32==311
pyreadline3==3.5.4
win_inet_pton==1.1.0
win32_setctime==1.2.0